pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vesicle_nrss"
description = "Single- and multi-vesicle NRSS morphology and simulation framework"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Dean DeLongchamp", email = "dean.delongchamp@nist.gov" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: Public Domain",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.23",
    "scipy>=1.9",
    "NRSS",
    "PyHyperScattering",
]
dynamic = ["version"]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x",
]
ray = [
    "ray",
    "ray_glados",
]
dev = [
    "pytest>=7.0",
]

[project.urls]
Homepage = "https://github.com/usnistgov/"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/vesicle_nrss/_version.py"
version_scheme = "no-guess-dev"
local_scheme = "no-local-version"
